+++ crates/nwc/src/options.rs
@@ -6,13 +6,24 @@
 
 #[cfg(not(target_arch = "wasm32"))]
 use std::net::SocketAddr;
+use std::time::Duration;
 
 use nostr_relay_pool::RelayOptions;
 
 /// NWC options
-#[derive(Debug, Clone, Default)]
+#[derive(Debug, Clone)]
 pub struct NostrWalletConnectOptions {
     pub(super) relay: RelayOptions,
+    pub(super) timeout: Duration,
+}
+
+impl Default for NostrWalletConnectOptions {
+    fn default() -> Self {
+        Self {
+            relay: RelayOptions::default(),
+            timeout: Duration::from_secs(10),
+        }
+    }
 }
 
 impl NostrWalletConnectOptions {
@@ -26,6 +37,13 @@ impl NostrWalletConnectOptions {
     pub fn proxy(self, proxy: Option<SocketAddr>) -> Self {
         Self {
             relay: self.relay.proxy(proxy),
+            ..self
         }
+    }
+
+    /// Set NWC requests timeout (default: 10 secs)
+    pub fn timeout(mut self, timeout: Duration) -> Self {
+        self.timeout = timeout;
+        self
     }
 }